New: debug option
Subprocess now uses the capture_ouput option whenever debug is
set to false. Each function that uses subprocess.run has a debug
argument. If the argument is set to True, the ouput capture is
set to false, allowing users to see what is going on. Debug is
set using a flag.